Refactor transforms and test fixed16.Float
Refactor transforms into their own package since they're resuable in
other places.

Add tests to fixed16.Float that ensure the conversion is exact and
relatively stable.

Float conversion is exact for values under 2^37,
which should cover most use cases. Getting to 2^38 is the highest
feasible but would be slightly slower.

For values higher than 2^37, there is a
relative conversion error of at most 2^-53, as the leftmost ten
digits are not guaranteed to be preserved. Further improvement
infeasible.